requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
mysql-connector-python==8.0.27
numpy==1.21.3
oauthlib==3.1.1
pandas==1.3.4
pkg-resources==0.0.0
protobuf==3.19.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tweepy==4.2.0
urllib3==1.26.7